basic switch

+++ crates/compiler/src/parse.rs
@@ -171,7 +171,7 @@ impl<'a, 'p> Parser<'a, 'p> {
                 self.start_subexpr();
                 self.start_subexpr();
                 self.pop();
-                let v = i64::from_le_bytes((value as u64).to_le_bytes());
+                let v = i64::from_le_bytes((value).to_le_bytes());
                 let v = self.expr(Expr::int(v));
                 let bits = self.expr(Expr::int(bit_count as i64));
                 let pair = self.expr(Expr::Tuple(vec![bits, v]));
